+++ src/java/org/apache/poi/hssf/record/formula/functions/Today.java
@@ -32,7 +32,7 @@ import org.apache.poi.hssf.usermodel.HSSFDateUtil;
  */
 public final class Today implements Function {
 
-	public ValueEval evaluate(ValueEval[] evals, int srcCellRow, short srcCellCol) {
+	public ValueEval evaluate(ValueEval[] evals, int srcCellRow, int srcCellCol) {
 		if (evals.length > 0) {
 			return ErrorEval.VALUE_INVALID;
 		}
